pyproject.toml
# pyproject.toml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "odocxify"
version = "0.1.0"
description = "An intelligent, AI-powered documentation generator for Python."
authors = [
    {name = "Shridhar Mandrekar", email = "shridharmandrekar118@gmail.com"},
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["documentation", "generator", "ai", "docstrings", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Documentation",
]

[project.urls]
Homepage = "https://Shridharrr.github.io/odoc/" # <-- UPDATE THIS
Repository = "https://github.com/Shridharrr/odoc" # <-- UPDATE THIS

[project.scripts]
odoc = "odocxify.main:app"

# --- Poetry-specific configuration remains below ---

[tool.poetry]
packages = [{include = "odocxify", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
jinja2 = "^3.1.6"
python-dotenv = "^1.1.1"
google-generativeai = "^0.8.5"
mkdocs = "^1.6.0"
mkdocs-material = "^9.0.0"
mkdocs-awesome-pages-plugin = "^2.9.1"

[tool.odoc]
exclude = [
    "tests/*",
    "ai_test.py"
]